refactor logics and specs, add logic and spec for create-training, remove-player-from-group and get-groups-for-coach, create helpers for create-training logic #13

+++ staff/sabrina-algaze/project/api/logic/deleteGroup.js
@@ -1,20 +1,26 @@
-import { validate, NotFoundError, RoleError, SystemError } from 'com'
+import { validate, NotFoundError, RoleError, SystemError, OwnershipError } from 'com'
 import { Group, User } from '../data/index.js'
 
-export const deleteGroup = (userId, groupId) => {
-    validate.userId(userId)
+export const deleteGroup = (coachId, groupId) => {
+    validate.userId(coachId)
     validate.groupId(groupId)
 
-    return User.findById(userId)
+    return User.findById(coachId)
         .catch(error => { throw new SystemError('mongo error') })
         .then(user => {
             if (!user) throw new NotFoundError('user not found')
             if (user.role !== 'coach') throw new RoleError('user role is not coach')
 
-            return Group.findByIdAndDelete(groupId)
+            return Group.findById(groupId)
                 .catch(error => { throw new SystemError('mongo error') })
                 .then(group => {
-                    if (!group) throw new NotFoundError('group not found')
+                    if (group.owner.toString() !== coachId) throw new OwnershipError('user is not the owner')
+
+                    return Group.findByIdAndDelete(groupId)
+                        .catch(error => { throw new SystemError('mongo error') })
+                        .then(group => {
+                            if (!group) throw new NotFoundError('group not found')
+                        })
                 })
         })
 }
